requirements.txt
Flask
openai
python-dotenv
Flask-SQLAlchemy
Flask-Session